os/bluestore: Cputrace measurements for core BlueStore functions
This work uses cputrace framework to introduce ready-to-use measurements
of selected core BlueStore functions. The list is not exhaustive and
in the future more functions might be covered by this mechanism.
Required WITH_CPUTRACE enabled for the project
and setting BLUESTORE_COMMON_CPUTRACE in BlueStore.h.

The feature is a developer tool indented to be used only when CPU related
optimizations are tested. The benefit is that PRs are ready to test
without creating side branches that include the measurements.

Signed-off-by: Adam Kupczyk <akupczyk@ibm.com>
Signed-off-by: Jaya Prakash <jayaprakash@ibm.com>
(cherry picked from commit 4ef5f3fd98330aa3c2f320827a24eaf60380ad1b)